Methods for loading points & edges moved to file loader module.
