+++ filehandler.h
@@ -2,7 +2,8 @@
 #define FILEHANDLER_H
 
 #include "errorinfo.h"
-#include "fileloader.h"
+#include "modeldata.h"
+#include <iostream>
 
 ErrorInfo file_handle_points(FILE *, PointVectorInfo *);
 ErrorInfo file_handle_edges(FILE *, EdgeVectorInfo *);
